# backend/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API proyectoUTM",
        "description": "<html><body>Esta es la documentación de la API para el sistema de venta de una carniceria.<p>Para poder usar esta API es necesario solicitar el token y con esto poder utilizar los servicios proporcionados.</p><br><p><b>Autenticación y token de API</b></p><p>Kevin Ivan Caballero Hernandez alumno de la Universidad Tecnológica de la Mixteca (UTM) proporciona el api-token. Una vez que haya creado y confirmado su cuenta.</p></body></html>",
        "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
    }
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
        "http"
        ],
    "consumes":[
        "application/json"
        ],
    "produces" :[
        "application/json"
        ],
    "paths":{
...
        "/api/users/validarUsuario/":{
            "post":
            {
                "summary" : "Validando un usuario",
                "operationId": "validarUsuario",
                "produces":[
                    "application/json"
                ],
                "parameters": [
                    {
                        "name" : "Usuario",
                        "in": "body",
                        "description" :"Verifica si el correo y contraseña existen",
                        "required" : true,
                        "schema" : {
                            "$ref" : "#/definitions/validarUsuario"
                        }
                    }
                ],

                "tags" : ["Usuarios"],
                
                "responses" : {
                    "200" :{
                        "description" : "Insertado con exito"
                    },
                    "403" : {
                        "description" : "Prohibido"
                    }
                }

            }
        },    
...
    },
    "definitions":{
...
        },"validarUsuario":{
            "type": "object",
            "properties":{
                "correo" :{
                    "type": "string",
                    "description" : "Correo del usuario"
                },
                "contra":{
                    "type": "string",
                    "description" : "contraseña del usuario"
                }

            }
        },
...
    }
}